commit logging

+++ NorthWind.Web/Controllers/Api/ProductApiController.cs
@@ -30,25 +30,63 @@ namespace NorthWind.Web.Controllers.Api
         public async Task<JsonResult>  GetByPage()
         {
             var productResponse = await _productService.GetProduct();
+            if(productResponse != null)
+            {
+            _logger.LogInformation("Load Prouducts");
             return Json(productResponse);
+            }
+            else
+            {
+                _logger.LogError("Load Prouducts failed");
+                return Json(new { message = "Load Prouducts failed" });
+            }
+
         }
 
         [HttpPost]
         public async Task  Create([FromBody] Product product)
         {
-            await _productService.InsertProduct(product);
+            if(product != null)
+            {
+                await _productService.InsertProduct(product);
+                _logger.LogInformation("Create Prouduct successfully");
+
+            }
+            else
+            {
+                _logger.LogError("Create Prouduct failed");
+            }
+           
         }
 
         [HttpPost]
         public async Task  Update([FromBody] Product product)
         {
-            await _productService.UpdateProduct(product);
+            if(product !=null)
+            {
+                await _productService.UpdateProduct(product);
+                _logger.LogInformation("Update Prouduct");
+            }
+            else
+            {
+                 _logger.LogError("Update Prouduct failed");
+            }
+
         }
 
         [HttpPost]
         public async Task  Delete([FromBody] int productId)
         {
-            await _productService.DeleteProduct(productId);
+            if(productId >0)
+            {
+                await _productService.DeleteProduct(productId);
+                _logger.LogInformation("Delete Prouduct");
+            }
+            else
+            {
+                _logger.LogError("Delete Prouduct failed");
+            }
+
         }
 
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
